requirements.txt
aiohttp>=3.8.0
deep_translator==1.8.3
langdetect==1.0.9
nextcord==2.6.0
nextcord_ext_menus==1.5.4
python-dotenv==0.20.0
requests==2.31.0
Telethon==1.26.0
